Improve the design of the data tables and rename class
I was criticized for the big data tables in the class. Unfortunately, it
cannot be done without: The mapping of the code points of the character
map is absolutely arbitrary, so listing is in this case the only option.
To improve readability, I tried to format them better. My first throw,
however, was not perfect yet. So I’ve improved that again.

Also, I renamed the class again because it still had the old C-style
name, which is uncommon in the Java environment. I will not rename it
again in the future, so as not to break the interface.